Simplified for long form only
Simplified the script, so that it puts everything in long form and has a new column for 'Surprise Index', which basically says how many of the same emotion expressions have come before this one since the last deviant. Note that the first of everyblock should be zero, but it's 1 (I guess that people also have the cue for what block it is, so they do somewhat expect the first deviant).